add support for morris.js
The hourly graph component relies on Morris.js. However, a dependency of Morris.js is Raphael.js.
When I tried to include the Raphael bower package and include it in my Brocfile, it broke the app.
After trying many different ways to fix this I concluded that there must be something outdated
with the Raphael bower package. The only workaround was to include the Raphael CDN in index.html.
This is hardly an ideal solution. I hope to gain a better understanding of how bower packages
work with EmberCLI and fix this issue in the future.