replace expanding tickets with new route
A lot of trouble has surrounded the expanding and collapsing of tickets.
This gets rid of that in favor of making the ticket details their own
route into which the ticket details are outletted. The next step is to
see if we can animate that outlet so it doesn't have to live at the
bottom of the ticket queue.