nest ticket list and details under queue
I noticed some really long load times as the route transitioned from a
ticket's details back to the queue. I realized this was because the
queue had to re-populate all of the tickets each time. Now, the index is
nested under the queue route (where tickets are loaded) as well as the
ticket details. Then transitioning back and forth is smooth since the
tickets never go away!